requirements.txt
aiosmtplib==2.0.1
anyio==3.6.2
blinker==1.5
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==39.0.2
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.95.0
fastapi-mail==1.2.6
h11==0.14.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
pycparser==2.21
pydantic==1.10.6
PyJWT==2.6.0
requests==2.28.2
sniffio==1.3.0
SQLAlchemy==2.0.7
starlette==0.26.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
